Add shared code dir and move all code to it
