+++ src/AI_Measurements/AI_Measurements.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net46</TargetFramework>
@@ -11,5 +11,9 @@
     <PackageReference Include="IllusionLibs.AIGirl.UnityEngine.UI" Version="2018.2.21.3" />
     <PackageReference Include="IllusionLibs.BepInEx" Version="5.4.8" />
     <PackageReference Include="IllusionModdingAPI.AIAPI" Version="1.20.2"></PackageReference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\Measurements.Core\**\*.cs" />
   </ItemGroup>
 </Project>
