+++ src/Measurements.Core/Gui/BustGui.cs
@@ -1,20 +1,20 @@
 ﻿using KKAPI.Maker;
 
-namespace AI_Measurements.Gui
+namespace Measurements.Gui
 {
-    internal class HipsGui : TextGui
+    internal class BustGui : TextGui
     {
         public override void Initialize(
             MakerCategory category,
             MeasurementsPlugin plugin,
             RegisterSubCategoriesEvent e
-        ) => InitializeInternal("Hips", category, plugin, e);
+        ) => InitializeInternal("Bust", category, plugin, e);
 
         public override void Update(MeasurementsData data, MeasurementsController controller)
         {
             SetText(controller.UseMetricUnits
-                ? $"{data.Hips:N0} cm"
-                : $"{data.Hips * FreedomRatio:N0}\"");
+                ? $"{data.Bust:N0} cm"
+                : $"{data.Bust * FreedomRatio:N0}\"");
         }
 
         protected override bool ShouldBeVisible()
