+++ src/Measurements.Core/Gui/IGui.cs
@@ -1,7 +1,7 @@
 ﻿using KKAPI.Maker;
 using KKAPI.Maker.UI;
 
-namespace AI_Measurements.Gui
+namespace Measurements.Gui
 {
     internal interface IGui<TControl> where TControl : BaseGuiEntry
     {
