+++ src/Measurements.Core/Gui/TextGui.cs
@@ -1,7 +1,7 @@
 ﻿using KKAPI.Maker;
 using KKAPI.Maker.UI;
 
-namespace AI_Measurements.Gui
+namespace Measurements.Gui
 {
     internal abstract class TextGui : IGui<MakerText>
     {
