+++ src/Measurements.Core/MeasurementsCalculator.cs
@@ -2,7 +2,7 @@
 using System;
 using UnityEngine;
 
-namespace AI_Measurements
+namespace Measurements
 {
     internal static class MeasurementsCalculator
     {
