+++ src/Measurements.Core/MeasurementsController.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-namespace AI_Measurements
+namespace Measurements
 {
     public class MeasurementsController : CharaCustomFunctionController
     {
